move ThemeStore under store/
