+++ electron-app/src/renderer/pages/configuration/configuration.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { Button, Divider, FormControlLabel, Paper, Radio, RadioGroup, Stack, styled, Typography } from '@mui/material'
 import { useResolve } from '../../hooks/container'
-import { ThemeStore, ThemeType } from '../../components/themeStore'
+import { ThemeStore, ThemeType } from '../../store/themeStore'
 import browserLogger from '../../logger/browserLogger'
 
 const Item = styled(Paper)(({ theme }) => ({
